pyproject.toml
[tool.poetry]
name = "vidyodl"
version = "0.1.0"
description = "Self-host your own video downloading API."
authors = ["MrKovar <kovarcody@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
ffmpeg-python = "^0.2.0"
pytube = "^15.0.0"
celery = "^5.3.4"
flower = "^2.0.1"
amqp = "^5.1.1"
pytest = "^7.4.2"
redis = "^5.0.1"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
pydantic-core = "2.10.1"
pydantic-settings = "^2.0.3"
asyncio = "^3.4.3"
pyamqp = "^0.1.0.7"
kafka-python = "^2.0.2"
asgiref = "^3.7.2"
black = "^23.10.1"
ruff = "^0.1.3"
isort = "^5.12.0"
debugpy = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
extend_skip = ["vidyodlvenv"]